Moved reusable asserts to their own class.
